Refactor identity steps UI (#831)
* refactor identity steps

* refactor identity steps

* refactor test and toast

* change classname

* refactor BEM

---------
