+++ app/components/identity-steps/step-five.js
@@ -1,42 +1,21 @@
 import Component from '@glimmer/component';
-import { tracked } from '@glimmer/tracking';
 import { action } from '@ember/object';
+import { tracked } from '@glimmer/tracking';
 import { inject as service } from '@ember/service';
-import { toastNotificationTimeoutOptions } from '../../constants/toast-notification';
 import { APPS } from '../../constants/urls';
-import checkURL from '../../utils/check-url';
+import { TOAST_OPTIONS } from '../../constants/toast-options';
 
 export default class IdentityStepsStepFiveComponent extends Component {
   @service toast;
-  @tracked isMouseOnTooltip = false;
-  @tracked profileURL = '';
-  @tracked nextButtonDisabled = true;
   @tracked isLoading = false;
 
-  @action openTooltipInfo() {
-    this.isMouseOnTooltip = true;
-  }
-
-  @action closeTooltipInfo() {
-    this.isMouseOnTooltip = false;
-  }
-
-  @action changeProfileURL(e) {
-    this.profileURL = e.target.value;
-    if (this.profileURL === '' || !checkURL(this.profileURL)) {
-      this.nextButtonDisabled = true;
-    } else {
-      this.nextButtonDisabled = false;
-    }
-  }
-
-  @action async handleEdit(e) {
+  @action async handleVerify(e) {
     e.preventDefault();
     this.isLoading = true;
+
     try {
-      const response = await fetch(`${APPS.API_BACKEND}/users/profileURL`, {
-        method: 'PATCH',
-        body: JSON.stringify({ profileURL: this.profileURL }),
+      const response = await fetch(`${APPS.API_BACKEND}/users/verify`, {
+        method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
@@ -44,27 +23,22 @@ export default class IdentityStepsStepFiveComponent extends Component {
       });
       if (response.ok) {
         this.toast.info(
-          'Updated profile URL!!',
-          '',
-          toastNotificationTimeoutOptions,
+          'Your request has been queued successfully',
+          'Info',
+          TOAST_OPTIONS,
         );
+        this.args.startHandler();
       } else {
-        this.toast.error(
-          'Something went wrong. Please check console errors.',
-          '',
-          toastNotificationTimeoutOptions,
-        );
+        throw new Error('Your request has not been queued');
       }
     } catch (error) {
-      console.error(error);
       this.toast.error(
         'Something went wrong. Please check console errors.',
-        '',
-        toastNotificationTimeoutOptions,
+        'Error',
+        TOAST_OPTIONS,
       );
     } finally {
       this.isLoading = false;
-      this.args.startHandler();
     }
   }
 }
