+++ app/components/identity-steps/step-one.hbs
@@ -1,11 +1,13 @@
 <h3 data-test-getting-started-heading class="heading__h3">Challenge Time !!</h3>
-<p data-test-getting-started-paragraph-1 class="p2-description">Thank you for providing all the details
-    Before joining the community we would want
+<p data-test-getting-started-paragraph-1 class="p2-description">
+    Thank you for providing all the details.
+    Before joining the community, we would want
     you to complete a small challenge which
     will also help you in setting up your
-    identity across Real Dev Squad
+    identity across Real Dev Squad.
 </p>
-<p data-test-getting-started-paragraph-2 class="p2-description">Please click proceed to know about
-    the task
+<p data-test-getting-started-paragraph-2 class="p2-description">
+    Please click proceed to know about
+    the task.
 </p>
 <Reusables::Button @text="Proceed" @variant='dark' @onClick={{@startHandler}} @test='identity-proceed' @type='button' />
