--- app/components/identity-steps/step-seven.hbs
@@ -1,86 +1,0 @@
-<div data-test-verification-page class='verification-page'>
-  {{#each this.profileStatuses as |currentStatus|}}
-    {{#if (eq this.login.userData.profileStatus currentStatus.status)}}
-      <div
-        data-test-verification-heading-icon
-        class='verification-page__heading-Icon'
-      >
-        <h3
-          data-test-verification-heading
-          class='verification-page__heading heading__h3'
-        >
-          {{currentStatus.heading}}
-        </h3>
-        <FaIcon
-          @icon={{currentStatus.icon}}
-          @size='lg'
-          class='{{currentStatus.icon}}-icon'
-        />
-      </div>
-      <div
-        data-test-verification-description-container
-        class='verification-page__description'
-      >
-        {{#if
-          (eq this.login.userData.profileStatus this.PROFILE_STATUS.PENDING)
-        }}
-          <p data-test-verification-description-1>
-            Refresh to Check Verification Status
-          </p>
-          <p data-test-verification-description-2>
-            Your Profile Service Linked with Real Dev Squad Service
-          </p>
-        {{else if (eq
-          this.login.userData.profileStatus this.PROFILE_STATUS.BLOCKED
-        )}}
-          <p data-test-verification-description-1>
-            Your previous
-            <b>Chaincode</b>
-            is Blocked.
-          </p>
-        {{else if (eq
-          this.login.userData.profileStatus this.PROFILE_STATUS.VERIFIED
-        )}}
-          <p data-test-verification-description-1>
-            <b>Congratulations!</b>
-            Your Profile Service is Verified.
-          </p>
-          <p data-test-verification-description-2>
-            Take the Next Step and Join Our Discord Server.
-          </p>
-        {{/if}}
-      </div>
-    {{/if}}
-  {{/each}}
-  <div class='verification-page__button'>
-    {{#if (eq this.login.userData.profileStatus this.PROFILE_STATUS.PENDING)}}
-      <Reusables::Button
-        @text='Refresh'
-        @variant='dark'
-        @onClick={{@handleRefresh}}
-        @test='refresh'
-        @type='button'
-      />
-    {{else if (eq
-      this.login.userData.profileStatus this.PROFILE_STATUS.BLOCKED
-    )}}
-      <Reusables::Button
-        @text='Verify Again'
-        @variant='dark'
-        @onClick={{@goToGenerateChaincodePage}}
-        @test='verify-again'
-        @type='button'
-      />
-    {{else if (eq
-      this.login.userData.profileStatus this.PROFILE_STATUS.VERIFIED
-    )}}
-      <Reusables::Button
-        @text='Next'
-        @variant='dark'
-        @onClick={{@startHandler}}
-        @test='next'
-        @type='button'
-      />
-    {{/if}}
-  </div>
-</div>
