+++ app/components/identity-steps/step-two.hbs
@@ -1,8 +1,13 @@
-    <h3 data-test-getting-started-heading class="heading__h3">Challenge Time !!</h3>
-    <p data-test-getting-started-paragraph-1 class="p2-description">To add/update your profile details, link
-        your profile service with Real Dev Squad service
-    </p>
-    <p data-test-getting-started-paragraph-2 class="p2-description">https://github.com/identity-service/instructions.md
-    </p>
-    <Reusables::Button @text="Next" @variant='dark' @onClick={{@startHandler}} @test='identity-next'
-        @type='button' />
+<h3 data-test-getting-started-heading class="heading__h3">Challenge Time !!</h3>
+<p data-test-getting-started-paragraph-1 class="p2-description">
+    To add/update your profile details, link your profile service with Real Dev Squad service
+</p>
+<a href="https://github.com/Real-Dev-Squad/sample-profile-service" target="_blank" rel="noopener noreferrer"
+   class="profile-service__link" data-test-profile-service-link>
+   https://github.com/Real-Dev-Squad/sample-profile-service
+</a>
+<p data-test-getting-started-paragraph-2 class="p2-description">
+    After exploring the sample profile service, proceed to the next step by clicking the button below.
+</p>
+<Reusables::Button @text="Next" @variant='dark' @onClick={{@startHandler}} @test='identity-next'
+    @type='button' />
