+++ tests/integration/components/identity-steps/step-two-test.js
@@ -7,7 +7,7 @@ module('Integration | Component | identity-steps/step-two', function (hooks) {
   setupRenderingTest(hooks);
 
   test('stepTwo renders', async function (assert) {
-    assert.expect(5);
+    assert.expect(6);
     let clicked = false;
     this.set('startHandler', () => {
       clicked = true;
@@ -16,6 +16,7 @@ module('Integration | Component | identity-steps/step-two', function (hooks) {
     await render(
       hbs`<IdentitySteps::StepTwo @startHandler={{this.startHandler}} />`,
     );
+
     assert
       .dom('[data-test-getting-started-heading]')
       .hasText('Challenge Time !!');
@@ -26,9 +27,18 @@ module('Integration | Component | identity-steps/step-two', function (hooks) {
       );
     assert
       .dom('[data-test-getting-started-paragraph-2]')
-      .hasText('https://github.com/identity-service/instructions.md');
-    assert.dom('[data-test-button=identity-next]').hasText('Next');
-    await click('[data-test-button=identity-next]');
+      .hasText(
+        'After exploring the sample profile service, proceed to the next step by clicking the button below.',
+      );
+
+    assert
+      .dom('[data-test-profile-service-link]')
+      .hasAttribute(
+        'href',
+        'https://github.com/Real-Dev-Squad/sample-profile-service',
+      );
+    assert.dom('[data-test-button="identity-next"]').hasText('Next');
+    await click('[data-test-button="identity-next"]');
     assert.true(clicked, 'click is called');
   });
 });
